feat(hex-point translation): make a hex's origin be relative to its center
I finally figured out how translating hexes to points, points to hexes, hexes to (view) pixels and
pixels to hexes. The breakthrough came with the realization that a hex's origin should be relative
to its center (instead of being relative to its top left point); redblobgames' `Hex#toPoint`
equivalent already returns the hex's center point. I just needed to subtract the hex's origin. When
a hex is rendered to a screen, only then the hex's center-originating point needs to be translated
to the DOM node's top-left-originating point.

hex origins are their center points by default with the value of {x:0, y:0}. `Hex#center` is removed
and `View#hexToPixel` returns a different value