change anchors of props `Point` and `Hex`
The two props `Point` and `Hex` took respective anchors which should have been used by factories of the same names instead.
As a result, internal links which were meant to be to those factories were actually to the props.
With "(prop)" added after the two props' headings, their anchors are now changed, and thus give way to the anchors of the factories.