feat(grid): more or less settled on Grid API
Grid contains several methods to iterate over it, each method returning a new grid with a different
"traverser" (a generator that yields hexes). This way a chain of lazy actions can be made, which can
all be invoked in sequence using the run() method. There's also a traverse() method that accepts one
or more "commands" (functions that return a traverser), this way the user can specify how a grid
should be traversed by composing commands that take the previous hex and yield 1 or more hexes. It's
beautiful 🥲