feat(hex,grid): traversers return hexes (as before), fix move() traverser, add neighborOf() function
The move() traverser didn't work properly, fixed it by converting to offset coordinates, update row
or col (depending on hex orientation) and then converting back to axial coordinates. Because of this
some new functions were needed: hexToOffset() and offsetToAxial(). Also added row and col getters
(which use hexToOffset()) on hex prototype for convenience. The functionality to get a neighboring
hex was moved from the move() traverser to a new neighborOf() grid function. This function also
supports ambiguous directions (North and South for pointy and East and West for flat hexes). But
because it was very awkward to create a hex (needed for the generic neighborOf() function) and then
returning only its coordinates in move(), I decided that all traversers should just receive and
return hexes (instead of receiving hexes but returning coordinates). For this I added a copyHex()
function (and delegating copy() method on hex prototype) that's able to copy the "cursor" hex a
traverser receives with new coordinates. This hex is then returned from traversers. Finally, both
copyHex() and createHex() now copy/create any custom props in their returning hex.