feat(grid): rectangle() traverser and grid method now accept opposing corners as arguments
The existing rectangleFromOpposingCorners() Grid method only worked for top left and bottom right
corners. First I fixed that it could handle any opposing corners (so also top right and bottom left)
in any order and then moved that to the rectangle traverser so that the grid method only needs to
delegate.