feat(grid): pass getHex() function to all traversers, this way they can use the cache
Also refactored the Grid's run() and traverse() methods: run() is simple again and traverse() is the
only method that sets the cache (although map() and neighborOf() should too). The impact on
performance is about 1% slower compared to a traverse that doesn't use caching at all. When multiple
traverse() calls are made the performance seems to increase up to ~25%(!)