refactor(grid): use regular Map as default hex "cache"
There was little added value in the CoordinatesCache. Also: using a noop cache (renamed to NoopMap)
would make it impossible to traverse a grid more than once. This is still the case, but I want to
remove the early return in traverse() when a NoopMap is used. Not sure yet how the API would look
though.