refactor(grid): rename several things and remove "magic" where first traversal sets state
Rules about what to do when traversing outside the grid doesn't belong in Grid's traverse() but
should be implemented in its other methods (like filter() and takeWhile() that need to be made yet).
A grid may be aware of state (in order to reuse hexes from that state), but how to use it should be
left to the user as much as possible. Also, not having this magic where the first traversal is
different from future traversals is needlessly complex.