feat(grid): pass cursor along with hexes in getPrevHexState() for a ~20% performance increase
When traversing a grid more than once the cursor was determined by doing a complete iteration of the
previous traverser in order to get the last hex. The next traverser continues where the previous
traverser left off. Now, the cursor is also returned when the previous traverser is called. This
increases performance with ~20% when traversing a 10k hex grid twice.