feat(grid): a grid can now be created with an optional traverser
A grid without traverser is pretty useless. So an optional traverser can now be passed as the 2nd
argument to the Grid constructor (or `Grid.of()`) and will be converted to internal "hex state".
Made the clone method private because I don't think people will use it and it requires an internal
function (GetHexState) is passed.