feat(grid): change how a grid's store is set
Instead of having to pass a store manually when creating a grid, the grid always makes a store. When
passed one or more traversers, the store is populated with the hexes that the traversers produce.
When passed a store, the grid's store is set to that store and it use those hexes to traverse over.

BREAKING CHANGE: Remove the of() static grid method, because it's a rather redundant as it does the
same as the constructor. Also update the inStore function to be used directly as a grid iterator
method callback (before change: `grid.each(inStore())`, after change: `grid.each(inStore)`). Remove
setStore() because it's ambiguous how the store should be set: hexes could be removed/added/updated
from the store.