Fix: Orientation being used as type and value
Orientation is a TS type, but in some places it's used as a string value: `orientation === Orientation.POINTY`

This causes issues, because, after the TS => JS compilation, there's no more TS type `Orientation`.
A better approach is to create an `Orientations` const that can be used in JS as well, as a value.
To get the type of such const, you can use `typeof Orientations`.

To say that some value needs to be one of the `Orientations` values, you can write: `ValueOf<typeof Orientations>`

This change also fixes the use of `'flat'`, `'FLAT'`, `'pointy'` and `'POINTY'` to allow only lowercase versions: `'flat'` and `'pointy'`