refactor(grid): remove cursor from Grid's internal hex state
Until now Grid kept track of where a traversal left off so that a next traversal would continue
there. This is a pretty ridiculous feature that added very little value and a lot of complexity 😅.
Note that when an array of traversers is passed to traverse() each consecutive traverser still
receives the cursor where the previous traverser left off (this is done in flatTraverse()).