feat(grid): remove move() alias of line() and update signature of line()
line() now accepts an object with options: `direction`, `length`, `at` and `start`. The `direction`
and `length` options are the same as before. The `at` option sets the cursor to the provided
coordinates, but doesn't include it. `start` does the same as `at` but *does* include those
coordinates. These `at` and `start` options will be added to most traversers.

BREAKING CHANGE: Remove move() alias of line(). line() now accepts an object with options: `direction`, `length`, `at` and `start`.