feat(grid): line() now accepts either "line as vector" options or "line between" options
The line() traverser can create a line of hexes starting at the given coordinates in a given
(compass) direction with a given length. Or it can create a line of hexes between start and end
coordinates. These end coordinates can be passed as either "until" or "through". Being similar to
"at" and "start", where "until" excludes the terminal coordinate and "through" includes the terminal
coordinate.