feat(grid): various changes, mainly traverse() and update() methods
With the transduce() function as the main way of traversing/updating hexes, Grid's traverse() and
update() are convenience methods that internally use transduce(). The added value of grid.traverse()
is that it returns hexes limited to those present in the grid. The added value of grid.update() is
that it always "merges" transformed hexes (optionally supplied by a traverser or iterable) into the
grid. It also automatically limits hexes to those present in the grid (passed transformers are only
called with existing hexes). Finally, the way a grid is created is improved. The constructor now
also accepts a grid (which will be cloned) or an iterable of hex coordinates.