feat: remove transducers
As it turns out, transducers don't perform a lot better compared to Array methods. They only perform
a little better (17% in a specific benchmark) when traversing a grid until the traverser "leaves the
grid". A transducer is able to shortcut the iteration. This seems a pretty rare use case and similar
functionality is implemented in the traverse() method (with the `bail` option). The examples are
updated to use the transducer-less API.